refactor: move js modules to modules directory
